testing moving getPath
