+++ src/app/actions/Report.js
@@ -1,4 +1,4 @@
-import urls from "./getPath";
+import urls from "../../../utils/getPath";
 
 export const createReport = async ({
   title,
